requirements.txt
dash==2.14.2
gunicorn==20.1.0
psycopg2-binary==2.9.9
pandas==2.1.4
plotly==5.18.0
dash-bootstrap-components==1.5.0
SQLAlchemy==2.0.29
